+++ metrics-core/src/main/java/com/yammer/metrics/util/DeadlockHealthCheck.java
@@ -1,9 +1,12 @@
-package com.yammer.metrics.core;
+package com.yammer.metrics.util;
+
+import com.yammer.metrics.core.HealthCheck;
+import com.yammer.metrics.core.VirtualMachineMetrics;
 
 import java.util.Set;
 
 /**
- * A {@link HealthCheck} implementation which returns a list of deadlocked threads, if any.
+ * A {@link com.yammer.metrics.core.HealthCheck} implementation which returns a list of deadlocked threads, if any.
  */
 public class DeadlockHealthCheck extends HealthCheck {
     private final VirtualMachineMetrics vm;