+++ CHANGELOG.md
@@ -31,6 +31,7 @@ v2.0.0-BETA19: TBD
   `#quantile` and `#quantiles`. This affects both `Histogram` and `Timer`.
 * Replaced `HealthCheck`'s abstract `name` method with a required constructor parameter.
 * `HealthCheck#check()` is now `protected`.
+* Moved `DeadlockHealthCheck` from `com.yammer.metrics.core` to `com.yammer.metrics.utils`.
 
 
 v2.0.0-BETA18: Dec 16 2011